pyproject.toml
[project]
name = "finsight"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # API
    "fastapi",
    "uvicorn",
    # Database
    "sqlalchemy[asyncio]",
    "alembic",
    "asyncpg",
    # Auth
    "pyjwt",
    "pwdlib[argon2]",
    "authlib",
    "httpx",
    # Validation & config
    "pydantic[email]",
    "pydantic-settings",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
